Correction of errors found by the mentor

+++ src/services/contacts.js
@@ -1,9 +1,8 @@
 import ContactCollection from '../db/models/Contact.js';
-
 import { sortList } from '../constants/index.js';
-
 import { calcPaginationData } from '../utils/calcPaginationData.js';
 
+// Отримання списку контактів із фільтрами, пагінацією та сортуванням
 export const getContacts = async ({
   page = 1,
   perPage = 10,
@@ -14,26 +13,24 @@ export const getContacts = async ({
   const skip = (page - 1) * perPage;
   const contactQuery = ContactCollection.find();
 
-  if(filters.userId) {
+  if (filters.userId) {
     contactQuery.where("userId").equals(filters.userId);
   }
-  
-  if(filters.contactType) {
+
+  if (filters.contactType) {
     contactQuery.where("contactType").equals(filters.contactType);
   }
-
-  // Фильтр по isFavourite
+  //Фільтр по isFavourite
   if (typeof filters.isFavourite === "boolean") {
     contactQuery.where("isFavourite").equals(filters.isFavourite);
   }
-    
+
   const totalItems = await ContactCollection.find().merge(contactQuery).countDocuments();
 
   const data = await contactQuery
     .skip(skip)
     .limit(perPage)
     .sort({ [sortBy]: sortOrder });
-  
 
   const paginationData = calcPaginationData({ page, perPage, totalItems });
 
@@ -44,16 +41,24 @@ export const getContacts = async ({
   };
 };
 
-export const getContactsById = (id) => ContactCollection.findOne({ _id: id });
+// Отримання контакту за id і userId
+export const getContactsById = (id, userId) =>
+  ContactCollection.findOne({ _id: id, userId });
 
+// Додавання нового контакту
 export const addContact = (payload) => ContactCollection.create(payload);
 
-export const updateContact = async (_id, payload, options = {}) => {
+// Оновлення контакту за id і userId
+export const updateContact = async (_id, userId, payload, options = {}) => {
   const { upsert = false } = options;
-  const rawResult = await ContactCollection.findOneAndUpdate({ _id }, payload, {
-    upsert,
-    includeResultMetadata: true,
-  });
+  const rawResult = await ContactCollection.findOneAndUpdate(
+    { _id, userId },
+    payload,
+    {
+      upsert,
+      includeResultMetadata: true,
+    }
+  );
 
   if (!rawResult || !rawResult.value) return null;
 
@@ -63,5 +68,6 @@ export const updateContact = async (_id, payload, options = {}) => {
   };
 };
 
-export const deleteContactById = (_id) =>
-  ContactCollection.findOneAndDelete({ _id });
+// Видалення контакту за id і userId
+export const deleteContactById = (_id, userId) =>
+  ContactCollection.findOneAndDelete({ _id, userId });
